+++ packages/pybamm-model-zoo/src/pybamm_model_zoo/spm_series_resistance/__init__.py
@@ -1,0 +1,3 @@
+from pybamm_model_zoo.spm_series_resistance.model import SPMSeriesResistance
+
+__all__ = ["SPMSeriesResistance"]